requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.6
PyMuPDF==1.25.5
pytesseract==0.3.13
Pillow==11.1.0
reportlab==4.3.1
pypdf==5.4.0
